requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
pydantic
requests
pytest
black
flake8